Add shop name to PawnTicketResponse and update ticket detail component for review functionality
Enhance the PawnTicketResponse DTO by adding a shopName field to include the shop's name in the response. Update the ticket detail component to display the shop name and implement a review section for redeemed tickets, allowing customers to submit and view reviews. Modify the corresponding service and HTML templates to support the new review features, improving user engagement and feedback collection.

+++ goldvault-backend/src/main/resources/db/migration/V7__create_review_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE shop_review (
     shop_id     BIGINT       NOT NULL,
     customer_id BIGINT       NOT NULL,
     ticket_id   BIGINT       NOT NULL,
-    rating      TINYINT      NOT NULL COMMENT '1 to 5 stars',
+    rating      INT          NOT NULL COMMENT '1 to 5 stars',
     comment     TEXT,
     status      VARCHAR(20)  NOT NULL DEFAULT 'VISIBLE' COMMENT 'VISIBLE or HIDDEN',
     created_at  DATETIME     NOT NULL DEFAULT CURRENT_TIMESTAMP,
@@ -13,6 +13,5 @@ CREATE TABLE shop_review (
     CONSTRAINT fk_review_customer FOREIGN KEY (customer_id) REFERENCES customer(id),
     CONSTRAINT fk_review_ticket   FOREIGN KEY (ticket_id)   REFERENCES pawn_ticket(id),
 
-    -- One review per ticket (customer can't review same ticket twice)
     CONSTRAINT uq_review_ticket UNIQUE (ticket_id)
 );
